move TranscriptSegment to its own file
